Rucia-루시아 Series/Tester-테스트기/requirements.txt
customtkinter==5.2.2
pillow
